chore: separate logic from bridge specific WIP
