+++ libs/coin-polkadot/src/bridge/signOperation.ts
@@ -1,12 +1,13 @@
 import { BigNumber } from "bignumber.js";
 import { Observable } from "rxjs";
-import { TypeRegistry } from "@polkadot/types";
-import { u8aConcat } from "@polkadot/util";
 import { FeeNotLoaded } from "@ledgerhq/errors";
 import type {
   PolkadotAccount,
+  PolkadotAddress,
   PolkadotOperation,
   PolkadotOperationExtra,
+  PolkadotSignature,
+  PolkadotSigner,
   Transaction,
 } from "../types";
 import type {
@@ -18,10 +19,10 @@ import type {
 } from "@ledgerhq/types-live";
 import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
-import { buildTransaction } from "./buildTransaction";
-import { calculateAmount, getNonce, isFirstBond } from "./logic";
 import { PolkadotAPI } from "../network";
-import { PolkadotAddress, PolkadotSignature, PolkadotSigner } from "../types";
+import { buildTransaction } from "../logic/buildTransaction";
+import { calculateAmount, getNonce, isFirstBond } from "../logic/logic";
+import { signExtrinsic } from "../logic/signTransaction";
 
 const MODE_TO_TYPE = {
   send: "OUT",
@@ -111,41 +112,6 @@ const buildOptimisticOperation = (
     extra,
   };
   return operation;
-};
-
-/**
- * Serialize a signed transaction in a format that can be submitted over the
- * Node RPC Interface from the signing payload and signature produced by the
- * remote signer.
- *
- * @param unsigned - The JSON representing the unsigned transaction.
- * @param signature - Signature of the signing payload produced by the remote signer.
- * @param registry - Registry used for constructing the payload.
- */
-export const signExtrinsic = async (
-  unsigned: Record<string, any>,
-  signature: any,
-  registry: TypeRegistry,
-): Promise<string> => {
-  const extrinsic = registry.createType("Extrinsic", unsigned, {
-    version: unsigned.version,
-  });
-  extrinsic.addSignature(unsigned.address, signature, unsigned as any);
-  return extrinsic.toHex();
-};
-
-/**
- * Sign Extrinsic with a fake signature (for fees estimation).
- *
- * @param unsigned - The JSON representing the unsigned transaction.
- * @param registry - Registry used for constructing the payload.
- */
-export const fakeSignExtrinsic = async (
-  unsigned: Record<string, any>,
-  registry: TypeRegistry,
-): Promise<string> => {
-  const fakeSignature = u8aConcat(new Uint8Array([1]), new Uint8Array(64).fill(0x42));
-  return signExtrinsic(unsigned, fakeSignature, registry);
 };
 
 /**
