+++ libs/coin-polkadot/src/logic/index.ts
@@ -1,0 +1,6 @@
+/**
+ * This directory is about pure coin logic.
+ * Its goal is to be as much as possible versatile code,
+ * and therefore can be used with "any Ledger product"
+ */
+
