+++ libs/coin-polkadot/src/network/index.ts
@@ -1,25 +1,3 @@
-/*
-
-// Alternative Implementations provided for example or future implementations
-
-export { getOperations } from "./subscan";
-
-export {
-  isElectionClosed,
-  isNewAccount,
-  isControllerAddress,
-  verifyValidatorAddresses,
-  getAccount,
-  getTransactionParams,
-  submitExtrinsic,
-  paymentInfo,
-  getValidators,
-  getStakingProgress,
-  getRegistry,
-  disconnect,
-} from "./websocket";
-
-*/
 import { NetworkRequestCall } from "@ledgerhq/coin-framework/network";
 import { getOperations as bisonGetOperations } from "./bisontrails";
 import {
