+++ libs/coin-polkadot/src/signer/index.ts
@@ -1,0 +1,4 @@
+/**
+ * This directory is the home for all types and logic based on Ledgers signer.
+ */
+
