+++ libs/ledger-live-common/src/families/polkadot/setup.ts
@@ -1,15 +1,15 @@
 // Goal of this file is to inject all necessary device/signer dependency to coin-modules
 
 import { createBridges } from "@ledgerhq/coin-polkadot/bridge/index";
+import type { Transaction } from "@ledgerhq/coin-polkadot/types/index";
 import makeCliTools from "@ledgerhq/coin-polkadot/test/cli";
 import type { CliTools } from "@ledgerhq/coin-polkadot/test/cli";
 import polkadotResolver from "@ledgerhq/coin-polkadot/bridge/hw-getAddress";
-import { Transaction } from "@ledgerhq/coin-polkadot/types/index";
 import Polkadot from "@ledgerhq/hw-app-polkadot";
 import Transport from "@ledgerhq/hw-transport";
 import { makeLRUCache } from "@ledgerhq/live-network/cache";
 import network from "@ledgerhq/live-network/network";
-import { Bridge } from "@ledgerhq/types-live";
+import type { Bridge } from "@ledgerhq/types-live";
 import { CreateSigner, createResolver, executeWithSigner } from "../../bridge/setup";
 import { Resolver } from "../../hw/getAddress/types";
